+++ src/sorting/insertion_sort.hpp
@@ -4,4 +4,4 @@
 using namespace std;
 
 void Insertion_Sort(int* arr, int n);
-void Insertion_Sort_Count(int* arr, int n, int& count);
+void Insertion_Sort_Count(int* arr, int n, long long& count);
