+++ src/sorting/merge_sort.hpp
@@ -4,4 +4,4 @@
 using namespace std;
 
 void Merge_Sort(int* arr, int n);
-void Merge_Sort_Count(int* arr, int n, int& count);
+void Merge_Sort_Count(int* arr, int n, long long& count);
