+++ src/sorting/quick_sort.hpp
@@ -5,4 +5,4 @@
 using namespace std;
 
 void Quick_Sort(int* arr, int n);
-void Quick_Sort_Count(int* arr, int n, int& count);
+void Quick_Sort_Count(int* arr, int n, long long& count);
